RESTWS-612 Merge inner wrapper composite key classes into one
two nested classes with almost identical implementation where used as a wrapper
class so two strings could be used as keys to the maps holding SearchHandlers.
first string holds the supported resource of the SearchHandler and second is
either the id or a parameter of the SearchHandler.

* renamed one of them to CompositeSearchHandlerKey
* first property stays supportedResource, second property is simply
additionalKeyProperty which is once populated with the SearchHandler id and
once with the parameter names.
* use instanceof in equals to check for null and if obj is of the same class
* added tests for equals/hashcode methods