RESTWS-720: API Docs generating a lot of reference errors (#342)
* RESTWS-720 Diagnose the fault as a nullPointerException and add a test to confirm.

Test simply calls the lower level method that throws the exception so
will pass once the fix is implemented.
Will also test for future regressions of the same issue.

* RESTWS-720 Fix nullPointerException

* RESTWS-720 Add test and change logic for create only SubResource definitions.

It was noted on refapp >=2.2.0 that OrderFulfillmentDetails was produce
a referr to its Get definition, an operation that wasn't supported by
the resource.

Logic was changed to remove subresource definitions in responses in line
with main resources. Test added to confirm and prevent regression.