RESTWS-817: Add search of concepts by class uuid
This allows the ConceptResource to search by a
class uuid and not require the name parameter to
be supplied. This allows returning of all the
concepts within a class to be returned in the
search query.

This patch will allow search route to work with
supplied "equals" and "fuzzy" search type but the
latter is essentially ignored. It also improves
the test for filtering concepts by name and by
class to test the "equals" search type as well.

https://issues.openmrs.org/browse/RESTWS-817